Fix delete status

+++ src/main/java/cz/upce/fei/dt/beckend/services/StatusService.java
@@ -3,7 +3,9 @@ package cz.upce.fei.dt.beckend.services;
 import com.vaadin.flow.data.provider.AbstractBackEndDataProvider;
 import com.vaadin.flow.data.provider.Query;
 import com.vaadin.flow.spring.data.VaadinSpringDataHelpers;
+import cz.upce.fei.dt.beckend.entities.Deadline;
 import cz.upce.fei.dt.beckend.entities.Status;
+import cz.upce.fei.dt.beckend.exceptions.UsedStatusException;
 import cz.upce.fei.dt.beckend.repositories.StatusRepository;
 import cz.upce.fei.dt.beckend.services.filters.StatusFilter;
 import cz.upce.fei.dt.beckend.services.specifications.StatusSpec;
@@ -12,12 +14,14 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.stream.Stream;
 
 @Service
 @AllArgsConstructor
 public class StatusService extends AbstractBackEndDataProvider<Status, StatusFilter> {
     private final StatusRepository statusRepository;
+    private final DeadlineService deadlineService;
 
     @Override
     public Stream<Status> fetchFromBackEnd(Query<Status, StatusFilter> query) {
@@ -42,8 +46,11 @@ public class StatusService extends AbstractBackEndDataProvider<Status, StatusFil
     }
 
     @Transactional
-    public void deleteStatus(Status status) {
+    public void deleteStatus(Status status) throws UsedStatusException {
+        List<Deadline> deadlines = deadlineService.findAllCurrentDeadlinesByStatusID(status.getId());
+        if (!deadlines.isEmpty())
+            throw new UsedStatusException(deadlines);
+
         statusRepository.delete(status);
     }
-
 }
